requirements.txt
python-telegram-bot==13.8.1
environs==9.3.5
vk-api==11.9.7